+++ tests/test_emptyDB.py
@@ -4,7 +4,7 @@ import sqlalchemy as sa
 import sys
 from pathlib import Path
 
-from weatherDB.db import models
+from weatherdb.db import models
 
 sys.path.insert(0, Path(__file__).parent.resolve().as_posix())
 from baseTest import BaseTestCases